pyproject.toml
[tool.poetry]
name = "py-spotify-unwrapped"
version = "0.1.0"
description = ""
authors = ["Martín do Río Rico <75792048+Meleagrista@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
spotipy = "^2.25.1"
notebook = "^7.4.2"
dotenv = "^0.9.9"
python-pptx = "^1.0.2"
pdf2image = "^1.17.0"
playwright = "^1.52.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
